better precision on the callback freq

+++ arcadatype.h
@@ -58,7 +58,7 @@ class Adafruit_Arcada_SPITFT {
   /**************************************************************************/
   virtual uint32_t variantReadButtons(void) = 0;
 
-  bool timerCallback(uint32_t freq, void (*callback)());
+  bool timerCallback(float freq, void (*callback)());
   void printf(const char *format, ...);
 
   // Filesystem stuff!
